requirements.txt
discord.py
discord.py[voice]
json
aiohttp